Adding missing libraries
Adding the missing embedded libraries that would cause some people to be missing their minimap icons.

If an addon loaded with both AceDB and AceAddon already loaded, the button would show.  If they loaded before these 2 addons loaded then the button would not show because it was not present yet.