1.5.0 (#57)
* Add currency icon

Add icon to currency on the list.

Close #36

* shorten amount

display shorten amount in currency list.  relates to issue #41

* begin list for icon selection

started to work on allowing user to select the minimap icon if they prefer - however, list significantly bogs down ui if automatically generated.  refernces issue #28

* fix world bosses

sync with master

* Add emmisary tracking

add tracking for emmisaries. issue #35

* Move emmisary and fix reset time

moved emmisary to display before currency.

fixed reset time to use generated reset as an offset to the current server time so it doesn't clear every time!

* Spelling mistake!

Emissary - not Emmisary

* emissary name

added space for display.

* change currency format

changed from 2 point precision to 1 point precision.

* Add event tracking

add event for player login.  restructure code so that this is more easily facilitated.

* change how events are registering.

added another event. work in progress.

* remove allCleared

remove so that data will show even when no world bosses or instances are locked to.  this will allow emmissary and currencies to show at all times.

* added weekly quests

added weekly/daily quests to be able to track blingtron and other quests (profession, class hall, etc).

* fix wow reset days

for the asian regions, it should be thursday.  correct this to calculate the reset on the proper date

* implement emissary fix

fix bug where emissary was not aligning properly across chars.  changed so that emissaryList is only loaded for the current player we are processing.

* implement issue #50

implement time left for instances when hovering.

* add reset time to world bosses.

* Config change - realm header

allow user to configure if the realm header hides or shows.

* show reset date

show reset date for blingtron.

* fix tooltip anchoring

fix so that tooltip anchors to the cell that is currently hovered over.  issue #50 closed

* fix weekly reset function

before was not calculating properly.  simplied so that it's easier to calculate the reset and has less complicated code.

* fixed occularus to correcct id

corrected occularus id from incorrect one to correct one.  this finishes all the argus world bosses.

* weekly lockout calc fix

fixed weekly lockout to calculate correctly.  Before was using the improper variable.  changed the variable to the correctg one and code is properly calcluating the reset.

* reconfigure how currency is displayed

remove hardcoding of displayText and move it into the minimap display.

also, added shortening as a configurable option in the config. This gets us close to the ability now to finalize the remaining issues for v1.5.0 release.

* fix for currency data

fix for nil error when converting currency from old format to the new format.

* Remove parenthesis

remove parenthesis from count and count/maximum.  should only be around bonus rolls

* issue #37

change so currency is color coded based on whether it is at it's max or not (if it has the capabiliity to max out)

white - no limit
red - limit reached
green - no limit reached.

* Add code to track more thank just blingtron

laying out the infrastructure to be able to track class hall cooldowns.  spells may prove a bit more difficult to track however...

* fix error

added and ( false ) instead of previous code.

* fix blingtron bug

fix bug for copyaccount wide.  we still need to copy, even if it's NOT completed. otherwise it never clears out the other characters.

* fix reset for instant quest.

figured out the proper calculation for the world quest resets.  works properly now.

* build currency table

build hardcoded table of all possible currencies.  add expansionLevel to allow grouping by expansion and a flag to show/hide from user (some we can get info for, but have been removed from game so no need to show)

* renamed weekly quest header

renamed weekly to repeatable as more of a catchall since they are not always daily or weekly (class hall quest for instance complete).

* create emissary table

create hardcoded emissary table since bonus is not reliable for finding completed emissary quests.

* Fix bug with emissary data

finally fixed bug so that emissary is properly setup.  Still a little flakey but only when you someone has not done the argus quests as the bonus objectives do not always show until server sync.

* add event - fix bug

add event for quest to rebuild data when quest changes.

also fix minor bug with the emissary data.

* add currency table.

add currency table so we can allow user to select which currency to hide or show.

* finalize fixes for currency selection

gearing up to close out #32 and #38.  in order to do this, #54 will need to be done for currency.  should be a quick fix to take care of then  finalize some regression testing and we're gtg!

* performance and moving

moving structs to helper class. minor bug fix with emissaries and housekeeping to finish off the currency piece.

* add final code for currency sorting and display

This should close issues #54 , #32  and #38 .  finishing testing and ready to release after!

* bug fix for emissary

showing 0/0 when tmeleft returns > 0 for completed quest but the objectives will return nil.

* Locale cleanup

Adding new values to the locale to properly display when translations are added.