track last login
add last logged in tracking.  since some api's are uavailable when logging out had to modify the initchardb() to not run certain blocks of code.

also, update last played with differential between the time it last ran and the current time (logging out).

closes issue #77

menu may need to be redone as it's slightly cluttered and for logged in chars, the data is hard to see when over the first column.